requirements.txt
Flask==2.3.2
Pillow==9.5.0
torch==2.0.0+cpu
torchvision==0.15.1+cpu
